qutely/dev-requirements.txt
freezegun
pywal
PyYAML
Jinja2
